Consolidate navigation and tracking state into /activity view
