+++ app/index.tsx
@@ -1,19 +1,20 @@
 import ChartView from "@/components/ChartView";
 import { RouteMode, useActiveRoute } from "@/hooks/useRoutes";
-import { router } from "expo-router";
+import { useTrackRecording } from "@/hooks/useTrackRecording";
+import { router, usePathname } from "expo-router";
 import { useEffect } from "react";
 
 export default function Index() {
+  const pathname = usePathname();
   const isNavigating = useActiveRoute((r) => r?.mode === RouteMode.Navigating);
+  const isRecording = useTrackRecording((s) => s.isRecording);
 
-  // On app load, if a navigating route was restored from persisted state,
-  // jump straight into the navigation screen. Only fires once per mount so
-  // we don't race with the normal start-navigation flow in RouteEditor.
+  // Re-present activity screen when dismissed back to index
   useEffect(() => {
-    if (isNavigating) {
-      router.navigate("/route/navigate");
-    }
-  }, [isNavigating]);
+    if (!isNavigating && !isRecording) return;
+
+    if (pathname === "/") router.navigate("/activity");
+  }, [pathname, isNavigating, isRecording]);
 
   return <ChartView />;
 }
